refactor: 이미지 업로드 로직 공통화
