+++ src/main/java/io/dnd/goyo/common/image/ImageType.java
@@ -1,0 +1,14 @@
+package io.dnd.goyo.common.image;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ImageType {
+    PLACE("place"),
+    REVIEW("review"),
+    USER("user");
+
+    private final String path;
+}
