+++ src/main/java/io/dnd/goyo/domain/place/controller/PlaceImageController.java
@@ -1,6 +1,7 @@
 package io.dnd.goyo.domain.place.controller;
 
-import io.dnd.goyo.domain.place.service.PlaceImageService;
+import io.dnd.goyo.common.image.ImageType;
+import io.dnd.goyo.common.image.ImageUploadService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.Map;
@@ -17,12 +18,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class PlaceImageController {
 
-    private final PlaceImageService placeImageService;
+    private final ImageUploadService imageUploadService;
 
     @Operation(summary = "이미지 업로드용 Presigned URL 발급", description = "장소 이미지를 업로드하기 위한 Presigned URL을 발급받습니다.")
     @GetMapping("/presigned-url")
     public ResponseEntity<Map<String, String>> getPresignedUrl(@RequestParam String filename) {
-        String presignedUrl = placeImageService.createPresignedUrl(filename);
+        String presignedUrl = imageUploadService.createPresignedUrl(ImageType.PLACE, filename);
         return ResponseEntity.ok(Map.of("url", presignedUrl));
     }
 }
